[utils]: Fix use of va_list in FormatStringArgs
`vsnprintf()` will internally change the state of the passed in va_list.
This results in incorrect logging as well as display output. Using
`va_copy()` to create a temporary copy of the va_list fixes this issue.

This commit also changes the iteration variables for certain loops to
match the correct size.